Revert "enhance(virtual-list): refactor virtual list (#1423)"
This reverts commit 6da4ddbb420795b4ad8da2fdab8207981acaf1ba.
