--- packages/web-vue/components/_components/virtual-list-v2/index.ts
@@ -1,3 +1,0 @@
-import VirtualList from './virtual-list.vue';
-
-export default VirtualList;
