--- packages/web-vue/components/_components/virtual-list-v2/style/index.less
@@ -1,7 +1,0 @@
-@import '../../../style/theme/index.less';
-
-@virtual-prefix-cls: ~'@{prefix}-virtual-list';
-
-.@{virtual-prefix-cls} {
-  overflow: auto;
-}
