+++ packages/web-vue/components/list/style/index.less
@@ -163,7 +163,7 @@
   }
 
   &-content {
-    flex: auto;
+    flex: 1;
 
     .@{prefix}-empty {
       display: flex;
